direcionamento de menu para métodos
